pyproject.toml
[project]
name = "llm-pdf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.3.3",
    "pymupdf>=1.27.2.3",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]

[project.scripts]
cli = "llm_pdf.cli:cli"
